+++ gulp/build.js
@@ -4,3 +4,8 @@ var gulp = require('gulp');
 var rump = require('../index');
 
 gulp.task(rump.taskName('build'), [rump.taskName('clean')]);
+
+gulp.task(rump.taskName('build:prod'), [
+  rump.taskName('prod:setup'),
+  rump.taskName('build')
+]);