+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### WIP
 - Add option to change task name prefix
+- Add `:prod` tasks for tasks with production
 
 #### 0.2.2
 - Add reconfigure option